fix migrations

+++ db/migrate/20190802182742_cleanup_tags.rb
@@ -1,5 +1,7 @@
 class CleanupTags < ActiveRecord::Migration[5.2]
   def change
-    ActsAsTaggableOn::Tag.destroy_all
+    if defined? ActsAsTaggableOn::Tag
+      ActsAsTaggableOn::Tag.destroy_all
+    end
   end
 end
